Move main crate to `crates/jwt-compact`
