+++ crates/jwt-compact/CHANGELOG.md
@@ -1,0 +1,1 @@
+../../CHANGELOG.md
