+++ crates/jwt-compact/README.md
@@ -1,0 +1,1 @@
+../../README.md
