+++ crates/jwt-compact/src/alg/eddsa_sodium.rs
@@ -1,24 +1,27 @@
-//! `EdDSA` algorithm implementation using the `ed25519-compact` crate.
+//! `EdDSA` algorithm implementation using the `exonum-crypto` crate.
 
 use core::num::NonZeroUsize;
 
-use ed25519_compact::{KeyPair, Noise, PublicKey, SecretKey, Seed, Signature};
-use rand_core::{CryptoRng, RngCore, TryRngCore};
+use anyhow::format_err;
+use exonum_crypto::{
+    PUBLIC_KEY_LENGTH, PublicKey, SEED_LENGTH, SIGNATURE_LENGTH, SecretKey, Seed, Signature,
+    gen_keypair_from_seed, sign, verify,
+};
 
 use crate::{
+    Algorithm, AlgorithmSignature, Renamed,
     alg::{SecretBytes, SigningKey, VerifyingKey},
     alloc::Cow,
     jwk::{JsonWebKey, JwkError, KeyType},
-    Algorithm, AlgorithmSignature, Renamed,
 };
 
 impl AlgorithmSignature for Signature {
-    const LENGTH: Option<NonZeroUsize> = NonZeroUsize::new(Signature::BYTES);
+    const LENGTH: Option<NonZeroUsize> = NonZeroUsize::new(SIGNATURE_LENGTH);
 
     fn try_from_slice(bytes: &[u8]) -> anyhow::Result<Self> {
-        let mut signature = [0_u8; Signature::BYTES];
-        signature.copy_from_slice(bytes);
-        Ok(Self::new(signature))
+        // There are no checks other than by signature length in `from_slice`,
+        // so the `unwrap()` below is safe.
+        Ok(Self::from_slice(bytes).unwrap())
     }
 
     fn as_bytes(&self) -> Cow<'_, [u8]> {
@@ -28,11 +31,19 @@ impl AlgorithmSignature for Signature {
 
 /// Integrity algorithm using digital signatures on the Ed25519 elliptic curve.
 ///
-/// The name of the algorithm is specified as `EdDSA` as per the [IANA registry].
+/// The name of the algorithm is specified as `EdDSA` as per [IANA registry].
 /// Use `with_specific_name()` to switch to non-standard `Ed25519`.
 ///
 /// [IANA registry]: https://www.iana.org/assignments/jose/jose.xhtml
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Hash)]
+#[cfg_attr(
+    docsrs,
+    doc(cfg(any(
+        feature = "exonum-crypto",
+        feature = "ed25519-dalek",
+        feature = "ed25519-compact"
+    )))
+)]
 pub struct Ed25519;
 
 impl Ed25519 {
@@ -40,14 +51,6 @@ impl Ed25519 {
     /// This is a non-standard name, but it is used in some apps.
     pub fn with_specific_name() -> Renamed<Self> {
         Renamed::new(Self, "Ed25519")
-    }
-
-    /// Generate a new key pair.
-    pub fn generate<R: CryptoRng + RngCore>(rng: &mut R) -> (SecretKey, PublicKey) {
-        let mut seed = [0_u8; Seed::BYTES];
-        rng.fill_bytes(&mut seed);
-        let keypair = KeyPair::from_seed(Seed::new(seed));
-        (keypair.sk, keypair.pk)
     }
 }
 
@@ -61,11 +64,7 @@ impl Algorithm for Ed25519 {
     }
 
     fn sign(&self, signing_key: &Self::SigningKey, message: &[u8]) -> Self::Signature {
-        let mut noise = [0_u8; Noise::BYTES];
-        rand_core::OsRng
-            .try_fill_bytes(&mut noise)
-            .expect("failed generating noise for Ed25519 signature");
-        signing_key.sign(message, Some(Noise::new(noise)))
+        sign(message, signing_key)
     }
 
     fn verify_signature(
@@ -74,13 +73,13 @@ impl Algorithm for Ed25519 {
         verifying_key: &Self::VerifyingKey,
         message: &[u8],
     ) -> bool {
-        verifying_key.verify(message, signature).is_ok()
+        verify(signature, message, verifying_key)
     }
 }
 
 impl VerifyingKey<Ed25519> for PublicKey {
     fn from_slice(raw: &[u8]) -> anyhow::Result<Self> {
-        Self::from_slice(raw).map_err(|err| anyhow::anyhow!(err))
+        Self::from_slice(raw).ok_or_else(|| format_err!("Invalid public key length"))
     }
 
     fn as_bytes(&self) -> Cow<'_, [u8]> {
@@ -90,15 +89,18 @@ impl VerifyingKey<Ed25519> for PublicKey {
 
 impl SigningKey<Ed25519> for SecretKey {
     fn from_slice(raw: &[u8]) -> anyhow::Result<Self> {
-        Self::from_slice(raw).map_err(|err| anyhow::anyhow!(err))
+        Self::from_slice(raw).ok_or_else(|| format_err!("Invalid secret key bytes"))
     }
 
     fn to_verifying_key(&self) -> PublicKey {
-        self.public_key()
+        // Slightly hacky. The backend does not expose functions for converting secret keys
+        // to public ones, and we don't want to use `KeyPair` instead of `SecretKey`
+        // for this single purpose.
+        PublicKey::from_slice(&self[SEED_LENGTH..]).unwrap()
     }
 
     fn as_bytes(&self) -> SecretBytes<'_> {
-        SecretBytes::borrowed(self.as_ref())
+        SecretBytes::borrowed(&self[..])
     }
 }
 
@@ -106,7 +108,7 @@ impl<'a> From<&'a PublicKey> for JsonWebKey<'a> {
     fn from(key: &'a PublicKey) -> JsonWebKey<'a> {
         JsonWebKey::KeyPair {
             curve: Cow::Borrowed("Ed25519"),
-            x: Cow::Borrowed(&key[..]),
+            x: Cow::Borrowed(key.as_ref()),
             secret: None,
         }
     }
@@ -119,10 +121,11 @@ impl TryFrom<&JsonWebKey<'_>> for PublicKey {
         let JsonWebKey::KeyPair { curve, x, .. } = jwk else {
             return Err(JwkError::key_type(jwk, KeyType::KeyPair));
         };
-        JsonWebKey::ensure_curve(curve, "Ed25519")?;
-        JsonWebKey::ensure_len("x", x, PublicKey::BYTES)?;
 
-        <PublicKey as VerifyingKey<_>>::from_slice(x).map_err(JwkError::custom)
+        JsonWebKey::ensure_curve(curve, "Ed25519")?;
+        JsonWebKey::ensure_len("x", x, PUBLIC_KEY_LENGTH)?;
+        Ok(PublicKey::from_slice(x).unwrap())
+        // ^ unlike some other impls, libsodium does not check public key validity on creation
     }
 }
 
@@ -130,8 +133,8 @@ impl<'a> From<&'a SecretKey> for JsonWebKey<'a> {
     fn from(key: &'a SecretKey) -> JsonWebKey<'a> {
         JsonWebKey::KeyPair {
             curve: Cow::Borrowed("Ed25519"),
-            x: Cow::Borrowed(&key[Seed::BYTES..]),
-            secret: Some(SecretBytes::borrowed(&key[..Seed::BYTES])),
+            x: Cow::Borrowed(&key[SEED_LENGTH..]),
+            secret: Some(SecretBytes::borrowed(&key[..SEED_LENGTH])),
         }
     }
 }
@@ -144,11 +147,11 @@ impl TryFrom<&JsonWebKey<'_>> for SecretKey {
             return Err(JwkError::key_type(jwk, KeyType::KeyPair));
         };
         let seed_bytes = secret.as_deref();
-        let seed_bytes = seed_bytes.ok_or_else(|| JwkError::NoField("d".into()))?;
-        JsonWebKey::ensure_len("d", seed_bytes, Seed::BYTES)?;
-        let seed_bytes = *<&[u8; Seed::BYTES]>::try_from(seed_bytes).unwrap();
+        let seed_bytes = seed_bytes.ok_or_else(|| JwkError::NoField("d".to_owned()))?;
 
-        let secret_key = KeyPair::from_seed(Seed::new(seed_bytes)).sk;
-        jwk.ensure_key_match(secret_key)
+        JsonWebKey::ensure_len("d", seed_bytes, SEED_LENGTH)?;
+        let seed = Seed::from_slice(seed_bytes).unwrap();
+        let (_, sk) = gen_keypair_from_seed(&seed);
+        jwk.ensure_key_match(sk)
     }
 }
