+++ crates/jwt-compact/src/error.rs
